add component state, top-level await support
