+++ examples/snowpack/astro/components/BaseLayout.astro
@@ -3,7 +3,7 @@ import Banner from './Banner.astro';
 import Nav from './Nav.astro';
 ---
 
-<Banner></Banner>
+<Banner />
 <Nav />
 
 <slot></slot>
