+++ examples/snowpack/astro/components/MainLayout.astro
@@ -1,9 +1,6 @@
 ---
 import BaseLayout from './BaseLayout.astro';
 import Menu from './Menu.astro';
-export function setup({ context }) {
-return {};
-}
 ---
 
 <BaseLayout>
