+++ examples/snowpack/astro/layouts/content-with-cover.astro
@@ -73,7 +73,7 @@ export let content: any;
             </h2>
             <div class="content-layout">
               <div class="content-body">
-                {content.body}
+                <slot></slot>
               </div>
             </div>
           </div>
