+++ src/@types/astro.ts
@@ -21,7 +21,7 @@ export interface JsxItem {
 
 export interface TransformResult {
   script: string;
-  props: string[];
+  imports: string[];
   items: JsxItem[];
 }
