+++ src/parser/index.ts
@@ -1,0 +1,1 @@
+export { default as parse } from './parse/index.js';
