+++ src/parser/parse/read/script.ts
