+++ src/parser/parse/state/text.ts
