+++ src/parser/parse/utils/entities.ts
