+++ src/parser/utils/namespaces.ts
