+++ src/parser/utils/trim.ts
